Ds_Virtual_space_micro/services/admin-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
slowapi==0.1.9
redis==5.2.1
python-jose[cryptography]==3.3.0
supabase==2.12.0
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.6
pydantic-settings==2.7.0